Add and implement getroles for userservice, delete,update and getstatuses for postservice.

+++ DataTier/mithus/src/main/java/dk/via/mithus/repositories/CostRepository.java
@@ -5,5 +5,5 @@ import org.springframework.data.jpa.repository.JpaRepository;
 import org.springframework.stereotype.Repository;
 
 @Repository
-public interface CostRepository extends JpaRepository<Cost, Integer> {
+public interface CostRepository extends JpaRepository<Cost, Long> {
 }
