+++ DataTier/mithus/src/main/java/dk/via/mithus/repositories/EnergyRatingRepository.java
@@ -5,5 +5,5 @@ import org.springframework.data.jpa.repository.JpaRepository;
 import org.springframework.stereotype.Repository;
 
 @Repository
-public interface EnergyRatingRepository extends JpaRepository<EnergyRating, Integer> {
+public interface EnergyRatingRepository extends JpaRepository<EnergyRating, Long> {
 }
