+++ DataTier/mithus/src/main/java/dk/via/mithus/repositories/HousingTypeRepository.java
@@ -5,5 +5,5 @@ import org.springframework.data.jpa.repository.JpaRepository;
 import org.springframework.stereotype.Repository;
 
 @Repository
-public interface HousingTypeRepository extends JpaRepository<HousingType, Integer> {
+public interface HousingTypeRepository extends JpaRepository<HousingType, Long> {
 }
