+++ DataTier/mithus/src/main/java/dk/via/mithus/repositories/PostStatusRepository.java
@@ -5,5 +5,5 @@ import org.springframework.data.jpa.repository.JpaRepository;
 import org.springframework.stereotype.Repository;
 
 @Repository
-public interface PostStatusRepository extends JpaRepository<PostStatus, Integer> {
+public interface PostStatusRepository extends JpaRepository<PostStatus, Long> {
 }
